+++ tapo/src/responses/device_info_result/default_state.rs
@@ -9,3 +9,22 @@ pub enum DefaultStateType {
     Custom,
     LastStates,
 }
+
+/// Default brightness state.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[cfg_attr(feature = "python", pyo3::prelude::pyclass(get_all))]
+#[allow(missing_docs)]
+pub struct DefaultBrightnessState {
+    pub r#type: DefaultStateType,
+    pub value: u8,
+}
+
+/// The type of the default power state.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+#[cfg_attr(feature = "python", pyo3::prelude::pyclass(get_all))]
+#[allow(missing_docs)]
+pub enum DefaultPowerType {
+    AlwaysOn,
+    LastStates,
+}