+++ CHANGELOG.md
@@ -6,6 +6,10 @@ file. This change log follows the conventions of
 
 ## [Rust Unreleased][Unreleased]
 
+## [Python Unreleased][Unreleased]
+
+## [Rust v0.7.7][v0.7.7] - 2024-01-13
+
 ### Changed
 
 - The `anyhow::anyhow!("Local hash does not match server hash")` error has been replaced with the more specific `tapo::TapoResponseError::InvalidCredentials` error.
@@ -15,7 +19,7 @@ file. This change log follows the conventions of
 - The `default_states` field that's part of the device info response has been changed for the L510, L520, and L610 devices to match the actual response from the device.
 - A handful of edge cases around the Klap Protocol that were causing panics have been fixed to return `tapo::TapoResponseError::SessionTimeout` or `tapo::TapoResponseError::InvalidResponse` errors instead.
 
-## [Python Unreleased][Unreleased]
+## [Python v0.1.4][py-v0.1.4] - 2024-01-13
 
 ### Changed
 
@@ -247,6 +251,8 @@ let device = ApiClient::new(ip_address, tapo_username, tapo_password)?
 ### Initial Release of Tapo
 
 [Unreleased]: https://github.com/mihai-dinculescu/tapo
+[v0.7.7]: https://github.com/mihai-dinculescu/tapo/tree/v0.7.7
+[py-v0.1.4]: https://github.com/mihai-dinculescu/tapo/tree/py-v0.1.4
 [v0.7.6]: https://github.com/mihai-dinculescu/tapo/tree/v0.7.6
 [v0.7.5]: https://github.com/mihai-dinculescu/tapo/tree/v0.7.5
 [py-v0.1.3]: https://github.com/mihai-dinculescu/tapo/tree/py-v0.1.3